Rename so_original_dst.rs to addr_info.rs.
Prepare for expanding the functionality of this module by renaming it.
